Never allow seeing checkpoints of unsupported version in log
Checkpoints of unsupported versions were allowed and a logtail
would be created of any other checkpoint seen. In some cases it still failed
because there was data after the last checkpoint (if in the same file). But if the
unsupported checkpoint was in a separate file it was just let through as
a valid tail.