Fixes a race in GBPTreeConsistencyChecker
Where the nbrRunning and the (now removed) lastOneSet
variables were checked in the wrong order and so the
"right-most in chain" state could be checked on incomplete data
and report mirage inconsistencies.

Now all shard state and nbrRunning is incremented before
starting any of the check tasks for those shards. The checking
threads can now just decrementAndGet nbrRunning to know if they're last.