Fix AppendAdvertisedAddressOnStateTransitionTest.shouldReturnAdvertisedAddress() ()
This test was failing because in some environments the paths uses "\" instead of "/", so the comparison with the static string doesn't work on those environments.

The fix for this is to use `.toString()` method from the `UnixDomainDomainSocketAddress` instead. This solution is not ideal since we are mimetizing the code implementation. However, this is needed for making those tests environment agnostic.