Fixes an issue removing rules from SchemaCache
During recovery (especially reverse recovery) there may be
occasions where there may be both an index and a constraint
with the same ID could exist at the same time.
Though the SchemaCache assumed that there could only be either or,
so when SchemaCache#removeSchemaRule(id) was called it would check
if there was a constraint with that ID and remove it, _else_
check if there was an index with that ID and remove it.
This made removing an index when both an index and a constraint
with the same ID impossible, and it would always choose to remove
the constraint.

This broken SchemaCache state would affect the IndexingService#init
and initialize wrong indexes. This in turn would affect how recovery
generated index updates, which affected block format recovery
specifically.