Changing the way summary is constructed for composite queries
There is a big issue with how the summary is currently constructed.
There is code that subscribes to driver's summary publisher when all
result stream records of a remote query are consumed, but there is
actually no code awaiting the summary's delivery. So if the driver
actually manages to publish the summary before the final aggregated
summary is produced is just a question of timing.

This PR simplifies the construction of aggregated summary for composite
queries and fixes the described issue.