Adding content type and term to envelope headers ()
Co-authored-by: Grant Lodge <6323995+thelonelyvulpes@users.noreply.github.com>

Update the envelope headers with two new fields: term (long) and contentType (byte).

The term is there because every raft log entry has a term and it is not uncommon for raft needing to look up terms for a specific index. It simplifies this a lot by storing the term in the header of the envelope.

The content type is simply there for filtering the log. We know that there will be readers that only interested in certain log entries.