feat: Refactor GenAI plugin to use `java.net.http.HttpClient` and port Apoc extended methods for external vector databases.
As base-work, this change introduces a new `HttpService` in the GenAI plugin that is used for both the existing procedure
(calls to OpenAI and friends for creating vector embeddings), and the new features introduced here (supporting the
storage of vector data together with identifiers pointing back to the graph) in external vector databases.

The reasons why `pom.xml`’s of other modules have been touched is that we want to use Testcontainers in a consistent
way, and we bring up several containers for databases in our integration tests (as a method preferred over to actually
calling an external cloud service).

Functionality of `genai.vector` encoding is otherwise unchanged.

The functional changes of this change are contained in the new namespace `genai.vector.external` in
`org.neo4j.genai.dbs.VectorDatabases`. This class makes use of `org.neo4j.service.Services` to load all
`org.neo4j.genai.dbs.ProviderResolver`. The default resovler makes use of the same mechanism again to load known
providers (those we ported from APOC extended). We have a second one in the testing code, who redirects all calls to
containers.

Apoc Extended does support
* ChromaDB
* Milvus
* Pinecone
* Qdrant
* Weaviate

and we ported all those, replicating the existing behaviour as much as possible. The biggest exception is not porting
the possibility to store any credentials or config, as this is out of scope and also in line with the existing GenAI
plugin story.

While apoc extended does have a separate namespace for all of these providers, this change does not. A method like

```
apoc.vectordb.qdrant.query(hostOrKey, collection, vector, filter, limit, #config)
```

becomes

```
genai.vector.db.query('https://xyz-example.eu-central.aws.cloud.qdrant.io', collection, vector, filter, limit,
#config)
```

And our provider resolver will transparently pick a provider that can deal with that URL. In this way we can assure that
this plugin can - especially when called from Aura - only reach known URL namespaces. In the above example,
`*.cloud.qdrant.io`.

Last but not least, the change reliefs users from knowing how to exactly to specify any auth header for the database
they want to call. In contrast to specify them with

```
:param configuration => ({Headers: {Authorization: 'Bearer your-api-key'}})
```

They just say

```
:param configuration => ({token: 'your-api-key'})
```

Procedures have been marked as `@Internal` so we can test our strategy around that feature.

Co-authored-by: Gerrit Meier <meistermeier@gmail.com>