Removal of reactive streams from composite code - phase 1
Since Java 21 introduced virtual threads, we can replace the reactive code with
good old blocking one without any loss of efficiency. The reactive code was
introduced, because there were concerns that if we used blocking driver API
in the composite (Formerly Fabric) code base, too many platform threads will be
created for workloads targeting many shards/constituents. The overall architecture
goal is to use blocking code and driver API and use Executors.newVirtualThreadPerTaskExecutor()
in the place where the fan out of query fragments to constituents happens.
The reactive code in composite (Formerly Fabric) code base has been a nightmare to
maintain over the past 4 years and has been a source of numerous concurrency issues.

This phase focuses on the local and remote executor. A new temporary utility
Blocking2RxResultAdapter is introduced. It is used for connecting StatementResult and
BlockingStatementResult together in the places where the old reactive code and
the new blocking code meet. It will be removed when the reactive code is gone.