Enhance CustomExtensionUtils createProcedure methods ()
The procedure will now return Stream.of(new Output(<string content>)),
rather than null. It is now also possible to control the procedure name,
in a predictable manner.

This allows us to remove a few of the "mixed" classloading that happens in the ReloadProcedureIT, and we take the opportunity to clean up in ReloadProcedureIT a bit.