Unflake LuceneIndexProviderTest#shouldHandleConcurrentUpdates()
It was using the lucene index populator in a way that never occurs
in the product and was adding data in a way that would make it
very hard to reason about the result.

The problem was that the sample data would sometimes be unexpected
(to the quite unexpected numbers that were asserted actually),
especially if there would be a call to "refresh" mid-populating.