StoreCopyCommand can better control max-off-heap-memory
Previously it could only control the size of the page cache
used for reading the source db. Rather than introducing another
option, the current option has been rebranded
(and gotten an additional name) to represent the total max off-heap
memory this store copy process is allowed to make use of, which is
what the user is interested in. This option also now defaults to
the same default calculation as e.g. that of the import command.

Internally 1/10th of this memory is given to the page cache
for reading the source db and 9/10ths to the importer of the target db.