Handle some recoverable last broken entry for enveloped logs
More handling for broken entry last in log for envelopes
Tries to differentiate between the cases where EnvelopeReadChannel has definitely seen something broken
and when it could be recoverable.
The EnvelopeReadChannel handles some more of this logic when it is a header being broken.

This is a slightly simplified solution where we assume it is a broken
last entry if we find some incomplete data and then no additional data
For example some data smaller than an envelope header is always assumed
to be an incomplete envelope. Same goes for when checksums mismatch
and there is no additonal data in the channel.