fix(SPD): plan remotebatchproperties on predicates outside connected components ()
Single component planner plans separate connected components and builds on top of them. This commit adds property accesses from "outside" the connected component, so that we can plan to fetch properties. This also ensure that getIndexValueBehaviours, if the predicate is using cached properties, it should rewrite the predicate to refer to these cached properties.