Don't enforce checksum chain by default when switching to new enveloped file in reading
There was a problem when starting a read at the end of a file:
When opening the EnvelopeReadChannel at a position, the needed segment is loaded but
the envelopes up to that point are only read if the position is not at the end. This should have been
okay except for the fact that loading in the next file tried to enforce the checksum chain even though
it had been marked as don't enforce. Now switching to the next file respects the
enforceChecksumChain variable.