StoreCopyCommand w/o node compaction properly updates .id file
When running store copy w/o compacting node ID space the potential holes
in between the imported nodes weren't written in the .id file.

This is now fixed, by using the same technique as block importer
does, which is to make a note about IDs being chosen externally
and rebuilding that particular .id file after import
(because keeping track of that during import is too expensive).