Fixing 2 bugs in handling standard GQL status codes in Composite database
1. EXPLAIN query on Composite does not return any status code.

There was a test that explicitly asserted this behaviour.

2. A wrong status code is returned for UNION quarries where the left part returns
results and the right one does not. The handling of standard GQL status codes was
implemented by taking the code from the last composite fragment. This obviously
does not work correctly in this case.

There was a test case for this scenario, but unfortunately the query it used was not
really a composite one. 'USE foo ... UNION USE foo ...' is not really a composite query
as the planner is smart enough to see that both parts of the union target the same
database and submits the entire union to the target database.