Implement KNL-019 and KNL-020
Also fix a bug where we got e very wrong error.
When creating a uniqueness constraint and then attempting to
drop the index belonging to that constraint in the same TX,
instead of "Unable to drop index: Index belongs to constraint",
we got "AssertionError: Expected 12 final commands, got 11 instead, with 0 skipped"

This is fixed by checking the TX state if the index to drop belongs to a newly
created constraint before calling `schemaRead.indexGetOwningUniquenessConstraintId`.