Change EnvelopeLogFiles MetadataCursor and use binary search for files ()
The metadata cursor was initially created for a different header structure and can therefor be simplified.

It is now just a RawIterator instead of RawCursor and does not need any closing.

This PR also removed the use of metadata cursor entirely when opening red channel and instead uses binary search for finding the correct file. Because the value from the binary search lives on disk I did not use any binary search library from collections or arrays, but perhaps there already some other implementation that can be used?

Lastly, this PR ensures that we handled pre-allocated files correctly. These will return a null header and would risk causing NPEs in the code