Fix some EnvelopedRecoveryCorruptedTransactionLogIT tests for windows
We are more permissive of broken last entry for envelope channels, if there
is ever less data than a header left we assume that it is a broken last entry.
On Windows the logs aren't preallocated so the logs weren't considered corrupted
with the amount of data that was appended before this fix. With preallocated logs
we would read the data and see that it was a corrupt header.