SPD: Entity population
This PR introduces a SPD-specific version of value population.
It is implemented as an execution engine decorator and it works like this:

A given number of rows coming from the wrapped execution engine is buffered.
The values are examined for nodes and relationship that need population
and IDs of those are collected. When the buffer is full, properties for the discovered
entities are fetched from the shards. After that the buffered rows are streamed to the consumer
during which the entities are populated with the fetched properties.

This version does not deal with memory tracking and query profiling.

The decorator implements QueryExecutionEngine interface, so this PR
also had to eliminate the direct use use of ExecutionEngine class. It does so
by introducing InternalQueryExecutionEngine interface which contains
the internal methods not present on QueryExecutionEngine.