Avoid using the configuration format setting in the LockVerificationFactory selection
To have the correct format, we need to make sure that the configuration is DatabaseConfig.
However, this is difficult to do correctly since selection is exposed over  Recovery in different places.
At the same time, we can work around this for now cheaper by making the selection a little later in the RecordStorageEngine constructor.