Fix RawTxLogCatchUpWriter failure case for envelope logs
With the enveloped log format it is not guaranteed that the start of a file is also the start of a
new appended batch (tx or chunk). The truncation point must be at the end of a completed appended batch
to be able to continue writing following pulled transactions correctly to the log.