Remove support for `file` scheme from `URLConnectionSeedProvider` ()
Co-authored-by: Nick Giles <nick.giles@neotechnology.com>
Co-authored-by: Balazs Lendvai <balazs.lendvai@neotechnology.com>

The file:// scheme was split into a separate provider: `FileSeedProvider`, but left on the existing `URLConnectionSeedProvider` for compatibility in the LTS. Now that we are post LTS, we can make people change configs.

The motivation is it enables people to say 'yes' to http / https URLs separately from their decision on file URLs, which is good for security.

Also:
InputStreamReadableChannel has an API where you pass in a byte array and a target number of bytes to put in it, and it returns when it has successfully done that. It calls an API on DataInputStream which (like InputStream) can read fewer bytes than the request, due to an IOException. The method needed changing to adapt these APIs to each other.