feat: Plan remote batch properties like interesting order ()
This commit introduces another requirement for additional properties so that a plan can pre-fetch properties from shards that can be used later. In particular a remoteBatchProperties before an expandAll so that we don't repeatedly fetch the same properties.
So BestResults now has the best overall plan, best sorted plan and best plan with additional properties.
This commit also updates PlanEventHorizon and SingleComponentPlanner to leverage additional properties.
PlanWithTail.planRhs also generates multiple plans now so that the planner can pick the best for apply plans.