Pass StoreIdGenerator to StorageEngineFactory ()
This removes StorageEngineFactory.resetMetadata. We instead provide a StoreIdGenerator to the StorageEngineFactory when it instantiates. This way the store can generate a StoreId based on the DatabaseOptions straight awa in enterprise dbs.

The reason is to simplify the RaftBootstrapper. In SPD we will create a new database type that only runs secondaries which would also have to do similar bootstrapping logic as the RaftBootstrapper, but with this change that will just work.