feat(SPD): replan RHS subquery ()
Ensure solveOptionalMatches will replan rhs if previouslyCachedProperties is non-empty. It will cache the context using the previously cached properties as the key, so that it doesn't replan for all LHS.
To avoid the overhead of caching in the non-SPD use-cases (where previouslyCachedProperties is always empty), the RHS in that case is pre-computed (using a lazy val).