SPD: Start node ID is not needed for reading rel. properties
Co-authored-by: Mattias Finné <mattias@neotechnology.com>

The SPD sharding function has been enhanced not to need start node ID for determining
the shard where a relationship is located. This can be used only for relationships that exist
in the store. The version of the sharding function requiring start node ID is still required
for creating relationships.