Add caching to `ContextBasedSystemDatabaseProvider`
Before for every `database()` call the system context was looked up in the context provider, which is a hashmap lookup - so it times some time.
Instead of that we can keep the resolved `DatabaseContext` and `GraphDatabaseAPI` in a cache and only reset that cache when the system database emits a `databaseCreate` event.