fix(planner): remove unnecessary contextualpropertyaccess computation ()
Currently we compute contextualPropertyAccess in both the head, and planTail. We also re-compute contextualPropertyAccess for each connected component. Ensure that we only compute contextualPropertyAccess in the planner head once. The property accesses from related components are only computed for Sharded databases, since we don't use the field in other database types.