PBAC allowsTraverseRelationship refactoring ()
In this PR I present my variation to the ReadSecurityPropertyProvider and interface of allowsTraverseRelationship (and others in future).
With ReadSecurityPropertyProvider user of the interface should know how to get list of security properties and then initialise above provider with those. And also expects instance of provider to memorise result after the first call.

Now, I introduce SelectedPropertiesProvider as argument for allowsTraverseRelationship. Which basically a function from PropertySelection to map of property values.
allowsTraverseRelationship takes care about supplying selection of properties of interest. And also it takes care to not call that function twice.