Don't include one transaction too much in reverse recovery for mvcc
Could be seen in restoreFromCheckpointInTheMiddleOfChunkedCommitedTransaction
with envelope profile on. The cursor position is the last read
after the tx which will match the checkpoint position exactly for
the last tx before the checkpoint. Seems to not have been an issue
for test without envelopes since rotation has happened just before the checkpoint.