Fix bug with relationship traversal cursor init
A relationship traversal cursor needs to reset the store cursor in init
if its node cursor is pointing to a node created in the transaction state.
This was not done in the init method that takes a neighbour node, that is
used by ExpandInto, so in rare cases when a traversal cursor is reused
in a cursor pool you can end up with an incorrect relationship match when
next is called on a store cursor that has not been reset.