Align ReversedEnvelopedCommandBatchCursor.position() with the other reversed cursors
The envelope version was returning the position after the batch (like forward cursors do) whereas the other
reversed cursors return the start position of the batch.
This fixes the problem where one batch too many was reversed in reverse recovery for mvcc when the oldestNotVisibleTransactionLogPosition and checkpoint position is not the same.