Add Vector types for type constraints ()
This change extends the TypeRepresentation with a new VectorType that tracks information about coordinate (numerical type) and dimension. Since we can not easily represent these types statically, we allow dynamic creation of the types. Each dynamically created type is interned in a type pool to avoid excessive creation of types.