Stop emitting warning when aggregation skips null
There has been a lot of complaints about this warning, and it has been decided we should
stop generating the warning. We still keep it in the code and just turn it off in configuration.
The reason for keeping it in the codebase is that the warning is specified in GQL and we might want
to put it back in a future GQL compliance mode.