Enforce alignment for SwappingBufferFactory
Since the backing file is shared between different array implementations, we have to make sure that all allocations are aligned to 8 bytes. Otherwise, an allocation of a smaller type, e.g., ByteArray of length 5, followed by a LongArray will make all accesses to the LongArray unaligned. This does not work for atomic operations.