Fix DataMatching
This utility did not really do what it promised with `dataMatchesEventually(cluster, databaseName)`. This method suppose to take the db representation of a database from every hosting server, and ensure that these are equal to each other at a given time. Instead of that, it reduced the task to compare the db representation of a database from the first server and then compare it repeatedly sequentially to every other representation. It would only compare with the next representation, once the previous comparison is successful. This would work if the first representation is stable, but that is not always the case.

For example: given we have three servers (S0,S1,S2), which need to eventually have the same data. We know that one of them has it - let's say (S2).
* start with: S0 txId: n; S1 txId: n; S2 txId: n+1;
* we compare S0 to S0 -> same -> success (while totally unnecessary)
* we compare S0 to S1 -> same -> success
* we compare S0 to S2 -> different -> repeat
* S0 manages to pull the extra transaction, so state is: S0 txId: n+1; S1 txId: n; S2 txId: n+1;
* we compare S0 to S2 -> same -> success

The method finishes while the 3 servers have different states (S1 could be behind). This is not at all what we wanted.

Other problems:
* the utility did await in two different ways
* it was used with `system`, but did not work for `system`, since `serversHostingAllOf` does not work for `system`