Better error message when lacking permissions to stop Neo4j
Previously, we would timeout on stop and give no clue that this
happened due to permissions issues. Now, we give a better error
message in this scenario. This is only implemented for Unix.

The error message reads:
"Neo4j (pid:42) process could not be stopped. Do you have the right permissions?"