SPD: Reintroducing long-running transactions on shards
Why bother with long-running transactions and not execute every query
in an autocommit transaction? From experience, we know that simple
RPC queries like fetching a single properties consume roughly the double
amount of CPU both on a shard and the main graph when executed in
an autocommit transaction compared to an already open long-running
transaction.