Wait for DbmsModelSnapshot for system db bookmarks
When waiting for a bookmark in system database the transaction id store of the system database was used.
This is almost correct, but the query routing stack will use components which are based on the dbms model (in enterprise edition) which gets updated after the transaction id store is already updated. In this PR we allow the `TransactionIdTracker` that for system database it can query the last transaction id against a `SystemLastTransactionIdProvider` interface - a component might be present in the dependency resolver. Since `DelegatingTopologyGraphDbmsModel` will now implement this new interface, in the enterprise edition the `TransactionIdTracker` will wait so until the DBMS model snapshot is also taken.