Treat all databases as unavailable, rather than not found, in routing stack ()
If a database is looked up in the context provider, and not found there, it has been treated as 'not found', which is a non-transient error. In fact to have reached the context lookup, you must have a database reference, implying that the database does exist to some extent, to the transient 'unavailable' is a better response.

In particular, this comes up because 'system' can now restart to change modes, and should never be considered 'not found'. However, the same is true for user databases if they are restarting etc.