fix(SPD Planner): order for parallel runtime and remoteBatchProperties
`RemoteBatchProperties(WithFilter)` is implemented under parallel runtime such that potentially several readers will feed out of a buffer, potentially invalidating a previously established order.

We therefore cannot rely on `RemoteBatchProperties` to propagate previous order but sort before any part of the plan the relies on the values arriving in order.