Refactor security graph object caches into one unified cache ()
Currently we have two caches for the security graph, one for privileges by role, and one for standard access mode. We also have a check for the user validity which runs at the start of every transaction and reads the database, which is not idea.

Refactor to have one cache, for Users, Role Privileges and AccessModes.

Co-authored-by: Phil Wright <95368282+phil198@users.noreply.github.com>