GlobalModule initializes its JobScheduler in init()
There was this weird thing where only the JobScheduler got a "manually"
init() call when created without really changing its life cycle's
state. Potentially it was to be able to schedule very long dbms
diagnostics logging during startup. Now the JobScheduler is
initialized in the GlobalModule's life's init() just like everything
else so that it's in sync with the GlobalModule's life.

This means that the system diagnostics dump happens slightly later,
with the seemingly added benefit of allowing to dump more information
in a clustered setup due to more diagnostics providers being available
at the time the dump is made.