DatabaseLayout no longer has methods for listing database files
This is because DatabaseLayout is pretty much context-less in terms
of file system, page cache and basically anything.

With the introduction of vector stores the files that comprise
a database is no longer static and the list of files that
"should exist" is more involved to come up with.

So instead now a file listing for a database is had either from
StorageEngine#listStorageFiles() or
StorageEngineFactory#listStorageFiles().

The use case for listing files is quite limited, save from some tooling
so this affects mostly tests, which uses these things much more.

This commit also removes the unused StoreFileMetadata wrapping
around paths in these file listings.