Introduce logFormatProvider instead of picking based on kernel version
The log format is still picked based on kernel version behind the
scenes for now, but this allows for a future when the log format
is not decided solely based on kernel version.
The current log format is inferred from the tx logs on start up,
or decided from the latest available version for new databases.
The current log format is possibly upgraded when applying
the upgrade transaction, with the new format present in the
upgrade command.

Some log rotations use an explicit log format because there are
some cluster scenarios where upgrade tx has not been applied at
the point of rotation, and therefore the logFormatProvider not
been updated. The protocol changes are not merged yet so
for now this is deciding log format on the receiver side based on
kernel version.