Assert assumption that key can be held in serialization buffer in IndexUpdateStorage ()
The index validators assume that we can serialize keys of size
`tree.keyValueSizeCap()` bytes. To validate that our code obeys this
contract, we add asserts to make sure that tests do not violate this
contract.

Executing with the asserts reveal that some tests used invalid buffer
sizes, and further that the validation constraint on the config value is
invalid.

Since this will become a problem as soon as we enable vector types,
we address these two concerns here.