Connect `SeedProcessor` to `DatabaseSeedProvider`s
But not for SPD checks. After this `CREATE DATABASE` and `recreateDatabase` will fail if no `DatabaseSeedProvider` can be found for the provided seed uri.
The `InternalDatabaseSeedProvider` interface has now a method for checking the provided uri and other seed options. The checking happens in two steps:
* First a `Checker` object can be acquired with the uri and options credentials, config and restore until
* Then the `Checker` object has a `checkFor` method which consumes the source database name and the flag for metadata script. `checkFor` returns a `Result` object which has
  * a method `resultCode` returning a `ResultCode` enum which is either `OK` or one of many `ERR_` fail values - latter needs to treated as the check has failed, the caller should throw
  * space for warnings - is not consumed at the moment
  * space for the metadata script - is not consumed at the moment

The implementations of `InternalDatabaseSeedProvider` at the moment return blindly `OK` in the `Checker`.