Remove `:Database` property `virtual`
This property was set to true always and only if the `:Database` node also had `:CompositeDatabase` label. The same information is stored twice, some components reacted on the property, some reacted on the label. Logically composite databases are only one type of virtual databases, but the codebase actually does not interpret this as such.