SPD: PropertyOperations for Execution Contexts ()
When RemoteTransactionContext was introduced, something important got
lost during the changes in the review cycle.

The original idea was that each Execution Context would have its own
instance of RemoteTransactionContext and transitively PropertyOperations
as PropertyOperations have a reference to RemoteTransactionContext.
As the situation stands before this PR, all Execution Contexts belonging
to one transaction actually use the same RemoteTransactionContext,
because they use the same PropertyOperations instance.

This is problematic for two reasons:

1. Many operations on RemoteTransactionContext are not thread safe for
   such use.
2. There is a performance problem as executing a query via
   RemoteTransactionContext is synchronised per location,
   so two Execution Contexts wanting to run a query on the same shard will
   block each other.

After the recent changes it does not seem really possible to have a
RemoteTransactionContext instance per Execution Context without a major
refactoring. So what I did instead is creating a
RemoteTransactionContext equivalent that is safe to be used concurrently
by multiple Execution Contexts and does not suffer from a performance
degradation when used in such a way. This new RemoteTransactionContext
equivalent is now used by Execution Contexts, but the old one is still
used outside Parallel runtime environment as it generally has better
performance.