Pass arguments and already cached properties to RemoteBatchPropertiesWithFilter ()
This PR updates `planSelection` to report solved on correct operator as
well as allows arguments and already cached properties to be inlined
into `RemoteBatchPropertiesWithFilter`.
It also adds a temporary case for `RemoteBatchPropertiesWithFilter` in
`InterpretedPipeMapper` so that we don't fall back to slotted runtime. A
[card](https://trello.com/c/DtMjJXMh/6381-allow-predicates-with-argument-references-in-shard-queries)
to do a correct fix for this has been created.