Feature/alt fix for spd reference ()
The general idea here is that we added the virtual database for a
reason, and it should be kinda like an alias, but kinda not. A good
enough cutoff is that if you make a transaction against a virtual
database, it should be resolved into the graph shard. Even though its
kinda possible to have a transaction against the virtual database and
this will work, feels .... not right. Makes more sense to me to never
have the virtual spd make it pass query router in this case.

Open to other better ideas but this is now attempt 3 at something
sensible.