Fix optional nullable slots (revert ) ()
The previous 'fix' did not really conform to the spirit of the logical
plan. The core problem is still that `availableSymbols` is a
`Set[LogicalVariable]` which does not necessarily have set semantics for
just variable names; perhaps the long term fix is to change
`availableSymbols` to operate on only names.

This PR works around the intended set intersection by including both LHS
and RHS variables for (Ordered)Union in `availableSlots`. It reverts the
previous attempt which delved into the physical plan's slot config