Use `SchemeFileSystemAbstraction.Factory` in the server process
Creating a `SchemeFileSystemAbstraction` is quite complicated as it requires a base `FileSystemAbstraction`, `Config`, `InternalLogProvider` and a `MemoryTracker`. Although there are defaults for all of them in the server process all of them should be coming from the `GlobalModule`. So instead letting all users creating their own a `Factory` is introduced, and the users can use that. With this we don't need to inject through all 4 components; some of them would be there, but especially the `MemoryTracker` would be just for this.