For MVCC BLOCK, preallocate max value size in gbptree for dense and huge ()
For MVCC to work neatly we need to have a way of being consistent with
value sizes. This PR introduces a way where we let all values be of the
same size when written to the tree, that being the maximum possible
value of what the tree can contain before we offload the value. This
works quite well for the HugeEntityStore but for DenseStore we can only
support 15 versions of the same entity, which is not ideal (the limit on
huge is 209). This space wasting only happens in the MVCC, normal block
is unaffected.

This is a temporary solution that we cannot rely on forever because it introduces great write amplification. We need it now because the current multi version solution only works for static sized values in GBPTrees.